Segmented parcel update
