+++ packages/react-app/src/hooks/index.js
@@ -7,3 +7,4 @@ export { default as useOnBlock } from "./OnBlock";
 export { default as usePoller } from "./Poller";
 export { default as useUserSigner } from "./UserSigner";
 export { default as useBurnerSigner } from "./BurnerSigner";
+export { default as useUpdateParcels } from "./UpdateParcels";
